Remove EpsRoleSelectionPage tests
